requirements.txt
antspyx==0.6.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
idna==3.11
imageio==2.35.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.7.5
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
psutil==7.1.0
pydicom==2.4.4
pyheatmap==0.1.12
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.3
requests==2.32.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
statsmodels==0.14.1
threadpoolctl==3.5.0
tifffile==2023.7.10
tzdata==2025.2
urllib3==2.2.3
webcolors==24.8.0
zipp==3.20.2
